requirements.txt
graphviz
ipython
matplotlib
neat-python==0.9
numpy
torch

-e .
